requirements.txt
python-dotenv 
pveagle 
pvrecorder 
customtkinter
vosk
numpy
PyQt6